+++ interfaces/kits/js/napi/src/is_last_work_time_out.cpp
@@ -20,10 +20,10 @@
 
 namespace OHOS {
 namespace WorkScheduler {
-const size_t WORK_ID_INDEX = 0;
-const size_t CALLBACK_INDEX = 1;
-const size_t IS_LAST_WORK_TIME_OUT_MIN_PARAMS = 1;
-const size_t IS_LAST_WORK_TIME_OUT_MAX_PARAMS = 2;
+const uint32_t WORK_ID_INDEX = 0;
+const uint32_t CALLBACK_INDEX = 1;
+const uint32_t IS_LAST_WORK_TIME_OUT_MIN_PARAMS = 1;
+const uint32_t IS_LAST_WORK_TIME_OUT_MAX_PARAMS = 2;
 
 struct IsLastWorkTimeOutParamsInfo {
     int32_t workId;
